Fixed: SourceOfRandomness#nextLong should call the "long" version

+++ core/src/main/java/com/pholser/junit/quickcheck/random/SourceOfRandomness.java
@@ -241,7 +241,7 @@ public class SourceOfRandomness {
         if (comparison == 0)
             return min;
 
-        return Ranges.choose(this, BigInteger.valueOf(min), BigInteger.valueOf(max)).longValue();
+        return Ranges.choose(this, min, max);
     }
 
     /**
